storage: add logic to allocate space on storage pool for 0-db
Each 0-db will run on top of a single disk.
When a new namespace is reserve inside a 0-db, we need to grow
the limit of the storage pool user by the 0-db.

Once the 0-db namespace is deleted, we can delete the namespace data
and then skring the limit of the storage pool.

The limit of the storage pool is used for capacity planning. We don't
allow to over provision storage on a storage pool