Replace original go PoC zoslog with C version
Since we potentially need to run lot of logger instance because we need
one per container, we need it to have the smallest memory footprint
possible.

After some test, the Go version was using ~3.6 MB of memory for a 3 MB
binary, with a basic Rust PoC, memory dropped down to ~1.6 MB.

This C version with static jansson and hiredis libraries bundled,
stripped binary size is 87K (138K fully static with musl).

On runtime, using glibc, shim-logs with 2 redis connections (one for
stdout and one for stderr) while transfering data, consume ~230 KB of
memory. The static musl versions seems to use even less but could not
get accurate values.

This only depends on hiredis and jansson external libs which are
compiled staticly by default with the Makefile.

In addition, there is a tools directory which contains a wrapper which
simulate a binary execution like it would be with containerd. This is
useful to test without actually run a container.