contd: do not wait forever on container delete
Request of container delete waits forever if the process cannot be
killed. The workflow tries to SIGTERM the running process then SIGKILL
after some try, but it never stop trying SIGKILL.

If process is in D or Z state, process won't be killed if stuck on
kernel call (occured with 'sync' on testnet). This make contd not
responding because stuck on deleting.

This fix will make contd responding again and not being stuck forever
but since an error is returned, provisiond will try to re-delete it
later.