Fix #1149: Sort out host ipv6 if br-pub is present (#1151)
In case br-pub is present in the host namespace, disable ipv6 on the
slave NIC. This is either the physical interface, or the zos bridge. In
case of the physical interface this is obviously fine. In case of the
zos bridge, it means host ipv6 traffic will need to pass additionally
over the br-pub and veth pair between br-pub and zos.

Signed-off-by: Lee Smet <lee.smet@hotmail.com>