Safer rrd calculations
The rrd algo now calculates the increase in value based on always
stored 'last' value. if no last value exist, we only take a note of
the last value. This means there is always an 'un-reported' consumption
for the first (window) but it's much safer. This means if there is a
huge gap in reports, the diff will not jump suddenly to 'total'
consumption