Safer upgrade managment (#1679)
Thanks to @delandtj we have to rethink our strategy to deploy
zos updates on the grid. This is separate the release creation
from the actual upgrade. While creating a release will build everything
but only on deploy workflow that has to be triggered manually the
release is going to be installed on the selected grid