Assign ipv6 and gw6 to vm via config
Instead of relying on slaac to give ipv6 to VMs now it's done
explicitly via cidata (cloud init data) provided to the vm

This means cloud-container now also disable router advertisment on
interfaces.

This also mean ipv6 is not provided unless requested in the workload.
This was a problem before because it was conflicting with other
configurable routes for ipv6