Make sure to only trigger event if ip actually change
I noticed on devnet that the node is continuesly trying to
re-register itself on the chain. This happened aftre we
add monitoring to changes to zos ip address.

There were some of the flows here:
- We handled changes to ALL ips on the host name space, not limited
  to the zos interface.
- For some reason, even with the device filter, the event keep happening
  even with `ip monitor address` command line, u can see events are
being triggered all the time, but no actual change (i assume it's slaac
since ip6 is what is included in the event)

To fix the previous errors, we did the following:
- Filter only changes on the zos bridge
- Make sure there is actually ip change before triggering the event